requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
backcall=0.2.0=pypi_0
bleach=3.2.1=pypi_0
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37he5f6b98_2
cffi=1.14.3=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
icu=67.1=he1b5a44_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=pypi_0
ipykernel=5.3.4=pypi_0
ipython=7.19.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuv=1.40.0=hd18ef5c_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=pypi_0
mistune=0.8.4=pypi_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.8=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.2=pypi_0
networkx=2.5=pypi_0
nodejs=14.14.0=h568c755_0
notebook=6.1.4=pypi_0
numpy=1.19.4=pypi_0
openssl=1.1.1h=h516909a_0
packaging=20.4=pypi_0
pandas=1.1.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=pypi_0
pip=20.2.4=py37h06a4308_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.8=pypi_0
protobuf=3.13.0=pypi_0
ptyprocess=0.6.0=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.2=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pypi_0
python_abi=3.7=1_cp37m
pytz=2020.4=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=19.0.2=pypi_0
readline=8.0=h7b6447c_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.5.0=pypi_0
setuptools=50.3.0=py37h06a4308_1
six=1.15.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboardx=2.1=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.10.1=pypi_0
timm=0.3.1=pypi_0
tk=8.6.10=hbc83047_0
torch=1.4.0=pypi_0
torchvision=0.5.0=pypi_0
tornado=6.1=pypi_0
tqdm=4.51.0=pypi_0
traitlets=5.0.5=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
